+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Docker image bugfix with not writing img digests to backup if container was down at backup execution
 - Skip Local bugfixes
 - Fixed bugs with Docker container not being restarted when fatalError during backup or transfer jobs
+- Custom output dir bugfixes
 
 ## [0.88.30] - 2025-1-28
 - Refactored all code into packages