+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.88.32] - 2025-1-29
+## [0.88.33] - 2025-1-29
 - Docker image bugfix with not writing img digests to backup if container was down at backup execution
 - Skip Local bugfixes
 - Fixed bugs with Docker container not being restarted when fatalError during backup or transfer jobs
 - Custom output dir bugfixes
 - Integrated gzip & tar functionality writers
+- Auto-copy ssh keys on remote transfer
 
 ## [0.88.30] - 2025-1-28
 - Refactored all code into packages